LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/flex-container-max-content-002.tentative-expected.txt

X

X

X

X X
X X

X X
X X

X X
X X
X

X

X

X X
X X

X X
X X

X X
X X
X

X

X

X X
X X

X X
X X

X X
X X
X

X

X

X X
X X

X X
X X

X X
X X

PASS .flex 1
FAIL .flex 2 assert_equals:
<div class="flex min" style="flex-flow: row nowrap" data-expected-width="45" data-expected-height="45">
  <div style="width: 25px; height: 25px">X</div>
</div>
height expected 45 but got 10
PASS .flex 3
PASS .flex 4
FAIL .flex 5 assert_equals:
<div class="flex min" style="flex-flow: row nowrap" data-expected-width="180" data-expected-height="45">
  <div data-expected-width="85" data-expected-height="35">X X</div>
  <div data-expected-width="85" data-expected-height="35">X X</div>
</div>
height expected 45 but got 10
PASS .flex 6
PASS .flex 7
FAIL .flex 8 assert_equals:
<div class="flex min" style="flex-flow: column nowrap" data-expected-width="45" data-expected-height="45">
  <div style="width: 25px; height: 25px">X</div>
</div>
height expected 45 but got 10
PASS .flex 9
PASS .flex 10
FAIL .flex 11 assert_equals:
<div class="flex min" style="flex-flow: column nowrap" data-expected-width="95" data-expected-height="80">
  <div data-expected-width="85" data-expected-height="35">X X</div>
  <div data-expected-width="85" data-expected-height="35">X X</div>
</div>
height expected 80 but got 10
PASS .flex 12
PASS .flex 13
FAIL .flex 14 assert_equals:
<div class="flex min" style="flex-flow: row wrap" data-expected-width="45" data-expected-height="45">
  <div style="width: 25px; height: 25px">X</div>
</div>
height expected 45 but got 10
PASS .flex 15
PASS .flex 16
FAIL .flex 17 assert_equals:
<div class="flex min" style="flex-flow: row wrap" data-expected-width="180" data-expected-height="45">
  <div data-expected-width="85" data-expected-height="35">X X</div>
  <div data-expected-width="85" data-expected-height="35">X X</div>
</div>
height expected 45 but got 10
PASS .flex 18
PASS .flex 19
FAIL .flex 20 assert_equals:
<div class="flex min" style="flex-flow: column wrap" data-expected-width="45" data-expected-height="45">
  <div style="width: 25px; height: 25px">X</div>
</div>
height expected 45 but got 10
PASS .flex 21
PASS .flex 22
FAIL .flex 23 assert_equals:
<div class="flex min" style="flex-flow: column wrap" data-expected-width="180" data-expected-height="45">
  <div data-expected-width="85" data-expected-height="35">X X</div>
  <div data-expected-width="85" data-expected-height="35">X X</div>
</div>
width expected 180 but got 95
PASS .flex 24
